+++ docs/changelog.rst
@@ -16,6 +16,7 @@ ChangeLog
 
     - Remove associated class discovery
     - Remove :class:`~factory.InfiniteIterator` and :func:`~factory.infinite_iterator`
+    - Remove :class:`~factory.CircularSubFactory`
     - Stop defaulting to Django's ``Foo.objects.create()`` when "creating" instances
     - Remove STRATEGY_*
     - Remove :meth:`~factory.Factory.set_building_function` / :meth:`~factory.Factory.set_creation_function`